Add DataService tests. Update old tests.

+++ step-release-vis/src/app/components/form/dataSubmissionForm.ts
@@ -73,6 +73,6 @@ export class DataSubmissionFormComponent implements OnInit {
   }
 
   arrayBufferToString(buf: ArrayBuffer): string {
-    return String.fromCharCode.apply(null, new Uint16Array(buf));
+    return String.fromCharCode.apply(null, new Uint8Array(buf));
   }
 }
